Created Alien Class.  Moved it to Space scene along with spawnAlien functionality
