+++ DoctorWizard/FireBall.swift
@@ -1,0 +1,9 @@
+//
+//  FireBall.swift
+//  DoctorWizard
+//
+//  Created by Rodrigo Carballo on 2/16/15.
+//  Copyright (c) 2015 codefellows. All rights reserved.
+//
+
+import Foundation
