+++ DoctorWizard/GameScene.swift
@@ -384,7 +384,7 @@ class GameScene: SKScene {
 
         case .Third:
             if !alienOn {
-                actionToSpawnAlien()
+                //actionToSpawnAlien()
                 println("Third scene on now")
             }
             
@@ -825,74 +825,7 @@ class GameScene: SKScene {
         return CGPoint(x: posX, y: posY)
     }
 
-    //MARK: SPAWN ALIENS ======================================================================
-    
-//    func spawnAlien() {
-//        let alien = SKSpriteNode(imageNamed: "alienspaceship")
-//        alien.name = "alienspaceship"
-//        let positionOnScreen = CGPoint(
-//            x: CGFloat.random(min: CGRectGetMinX(playableRect) + alien.frame.size.width,
-//                max: CGRectGetMaxX(playableRect)),
-//            y: size.height)
-//        alien.position = backgroundLayer.convertPoint(positionOnScreen, fromNode: self)
-//        alien.setScale(1)
-//        alien.zPosition = 0
-//        backgroundLayer.addChild(alien)
-//        var randomXPosition = CGFloat.random(min: 0, max: size.width)
-//        var randomYPosition = CGFloat.random(min: 0, max: size.height)
-//        let appear = SKAction.scaleTo(1, duration: 2.0)
-//        let moveToY = backgroundLayer.convertPoint(CGPoint(x: 0, y: randomYPosition), fromNode: self)
-//        let moveToX = backgroundLayer.convertPoint(CGPoint(x: randomXPosition, y: 0), fromNode: self)
-////        let actionMoveYDown =
-////        SKAction.moveToY(0, duration: 2.0)
-////        let actionMoveX =
-////        SKAction.moveToX(randomXPosition, duration: 0.5)
-////        let actionMoveYUp =
-//        
-//        let actionMoveY = SKAction.moveTo(moveToY, duration: 2.0)
-//        let actionMoveX = SKAction.moveTo(moveToX, duration: 1.0)
-//        SKAction.moveToY(size.height - alien.frame.height / 2, duration: 4.0)
-//
-//        let actionRemove = SKAction.removeFromParent()
-//        alien.runAction(SKAction.sequence([appear, actionMoveX, actionMoveY, actionRemove]))}
-    
-    
-    func spawnAlien() {
-        let alien = SKSpriteNode(imageNamed: "Spaceship")
-        alien.name = "alienspaceship"
-        alien.position = randomSpawnPoint()
-        alien.zRotation = CGFloat.random(min: 0, max: 90)
-        var directions : [SKAction] = []
-        var points : [CGPoint] = []
-        points.append(self.dude.position)
-
-        for i in 1...25 {
-            let randx = CGFloat.random(min: -250, max: 250)
-            let randy = CGFloat.random(min: -250, max: 250)
-            let point = CGPoint(x: points[i-1].x + randx, y: points[i-1].y + randy)
-
-            let action = SKAction.moveTo(point, duration: NSTimeInterval(0.1))
-            points.append(point)
-            directions.append(action)
-        }
-        
-        let remove = SKAction.removeFromParent()
-        directions.append(remove)
-        
-        self.backgroundLayer.addChild(alien)
-
-        
-        let move = SKAction.moveTo(self.dude.position, duration: 3)
-        let wait = SKAction.waitForDuration(1)
-        alien.runAction(SKAction.sequence([move,wait, SKAction.sequence(directions)]))
-        
-    }
-    
-
-    
-    
-    
-    
+ 
     //MARK: COLLISIONS ==========================================================================
     
     func dudeHitObject(enemy: SKSpriteNode) {
@@ -1278,15 +1211,7 @@ class GameScene: SKScene {
         println("FireBall on scene on now")
     }
     
-    func actionToSpawnAlien() {
-        alienOn = true
-        runAction(SKAction.repeatActionForever(
-            SKAction.sequence([SKAction.runBlock(spawnAlien),
-                SKAction.waitForDuration(7)])))
-        println("Alien on scene on now")
-    }
-    
-    
+        
     //MARK: start acceleromiter updates
